pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crosscompute-macros"
version = "0.0.1"
dependencies = []
requires-python = ">= 3.10"
authors = [
    {name = "Roy Hyunjin Han", email = "rhh@crosscompute.com"},
]
maintainers = []
description = "Reusable classes and functions needed by crosscompute packages"
readme = "README.md"
keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

[project.optional-dependencies]
disk = ["aiofiles >= 24.1.0"]
iterable = []
jinja = ["jinja2"]
log = []
markdown = ["markdown2"]
security = ["argon2-cffi"]
sqlalchemy = ["sqlalchemy"]
starlette = ["starlette"]
web = ["aiofiles", "aiohttp"]
yaml = ["ruamel.yaml", "ruamel.yaml.bytes"]
development = ["pre-commit"]
test = ["aiofiles", "pytest-asyncio", "pytest-cov", "pytest-xdist"]
all = [
    "aiofiles",
    "aiohttp",
    "argon2-cffi",
    "jinja2",
    "markdown2",
    "pre-commit",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-xdist",
    "ruamel.yaml",
    "sqlalchemy",
    "starlette",
]

[project.urls]
# Homepage = ""
# Documentation = ""
Repository = "https://github.com/crosscompute/crosscompute-macros"
Issues = "https://github.com/crosscompute/crosscompute-macros/issues"
# Changelog = ""

[tool.hatch.build]
packages = ["sources/crosscompute_macros"]
